// src/shared/types/themes.ts
export interface ColorTypes {
  MAIN: {
    COMMON: string;
    DARKER: string;
    LIGHTER: string;
    SUBTLE: string;
  };
  ERROR: {
    COMMON: string;
    DARKER: string;
    LIGHTER: string;
    SUBTLE: string;
  };
  WARNING: {
    COMMON: string;
    DARKER: string;
    LIGHTER: string;
    SUBTLE: string;
  };
  INFO: {
    COMMON: string;
    DARKER: string;
    LIGHTER: string;
    SUBTLE: string;
  };
  SUCCESS: {
    COMMON: string;
    DARKER: string;
    LIGHTER: string;
    SUBTLE: string;
  };
  DARK: {
    DARK1: string;
    DARK2: string;
    DARK3: string;
    DARK4: string;
  };
  LIGHT: {
    LIGHT1: string;
    LIGHT2: string;
    LIGHT3: string;
    LIGHT4: string;
  };
  LINEAR: {
    LINEAR1: string;
    LINEAR2: string;
  };
}

export interface TypographyTypes {
  HEAD: {
    600: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    500: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    400: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    300: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    200: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    100: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
  };
  BODY: {
    400: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    300: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    200: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    100: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
  };
  DISPLAY: {
    200: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
    100: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
  };
  LEADPARAGRAPH: {
    100: {
      fontSize: string;
      fontWeight: string;
      lineHeight: string;
    };
  };
}